requirements.txt
requests
sqlalchemy
mysqlclient
python-telegram-bot
asyncio
